# Film Manager Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  version: 1.0.0
  description: REST APIs that alllows to interact with the Film Manager Service
...
paths:
...
  "/api/films":
    post:
      tags: 
      - "api/films"
      summary: Used to add a new film to the owner list
      description: An authenticated user can create a film that is added to it's own list
      security:
        - cookieAuth: []
      operationId: createFilm
      requestBody:
        description: Represents the film that is going to be created
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
            examples:
              example1:
                value:
                  title: Inside Out 2
                  owner: 1
                  private: true
                  watchDate: "2024-11-27"
                  rating : 9
                  favorite: false
              example2:
                value:
                  title: Inception
                  owner: 1
                  private: false
        required: true
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
              examples:
                example1:
                  value:
                    id: 10
                    title: Inside Out 2
                    owner: 1
                    private: true
                    watchDate: "2024-11-27"
                    rating : 9
                    favorite: false
                    self: /api/films/public/10
                example2:
                  value:
                    id: 11
                    title: Inception,
                    owner: 1
                    private: false
                    reviews: /api/films/public/11/reviews
                    self: /api/films/public/11
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
    get:
      tags: 
      - "api/films"
      summary: Retrive all the films created by the authenticated user.
      description: An authenticated user can create all the films that he created.  A pagination machanism is used in order to limit the size of the messages.
      security:
        - cookieAuth: []
      operationId: getOwnedFilms
      parameters:
        - in: query
          name: pageNo                  
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example:
                {totalPages: 3,
                currentPage: 1,
                films: [
                    {
                      id: 10,
                      title: Inside Out 2,
                      owner: 1,
                      private: true,
                      watchDate: "2024-11-27",
                      rating : 9,
                      favorite: false,
                      self: /api/films/public/10
                    },
                    {
                        id: 11,
                        title: Inception,
                        owner: 1,
                        private: false,
                        reviews: /api/films/public/11/reviews,
                        self: /api/films/public/11
                    }
                ],
                next: /api/films?pageNo=2
                }
            


        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  "/api/films/private":
    get:
      tags: 
      - "api/films/private"
      security:
        - cookieAuth: []
      operationId: getPrivateFilms
      summary: retrieve the private films of the user.
      description: A logged-in user can retrieve his private films. A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requeted page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example: 
                {
                  "totalPages": 1,
                  "currentPage": 1,
                  "films": [
                      {
                          "id": 1,
                          "title": "Your Name",
                          "owner": 1,
                          "private": true,
                          "watchDate": "2021-10-03",
                          "rating": 9,
                          "self": "/api/films/private/1"
                      },
                      {
                          "id": 4,
                          "title": "Weathering with You",
                          "owner": 1,
                          "private": true,
                          "self": "/api/films/private/4"
                      }
                  ],
                  "next": null
              }
              
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  "/api/films/private/{filmId}":
    get:
      tags: 
      - "api/films/private/{filmId}"
      security:
        - cookieAuth: []
      operationId: getSinglePrivateFilm
      summary: retrieve a specific private film of the user.
      description: A logged-in user can retrieve a specific private film that he owns.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
              example: 
                id: 10
                title: Inside Out 2
                owner: 1
                private: true
                watchDate: "2024-11-27"
                rating : 9
                favorite: false
                self: /api/films/public/10

        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    put:
      tags: 
      - "api/films/private/{filmId}"
      security:
        - cookieAuth: []
      operationId: updateSinglePrivateFilm
      summary: Update a specific private film of the user.
      description: A logged-in user can update a specific private film that he owns except for it's visibility.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      requestBody:
        description: The film object that must be replaced to the old one
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
            example:
              title: Inside Out 2
              owner: 1
              private: true
              watchDate: "2024-11-27"
              rating : 10
              favorite: true
        required: true
      responses:
        "200":
          description: Succesfully Updated           
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict (The request is not allowed, like a change of visibility)
        "500":
          description: Internal Server Error

    delete:
      tags: 
      - "api/films/private/{filmId}"
      security:
        - cookieAuth: []
      operationId: deleteSinglePrivateFilm
      summary: Delete a specific private film of the user.
      description: A logged-in user can delete a specific private film that he owns.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      responses:
        "204":
          description: Successfully Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
  "/api/films/public":
    get:
      tags: 
      - "api/films/public"
      operationId: getPublicFilms
      summary: retrieve a list of public films.
      description: A generic user can retrieve a list of public films added by other users. 
        A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example: 
                {
                  "totalPages": 3,
                  "currentPage": 1,
                  "films": [
                      {
                          "id": 2,
                          "title": "Heaven's Feel",
                          "owner": 2,
                          "private": false,
                          "reviews": "/api/films/public/2/reviews",
                          "self": "/api/films/public/2"
                      },
                      {
                          "id": 3,
                          "title": "You Can (Not) Redo",
                          "owner": 1,
                          "private": false,
                          "reviews": "/api/films/public/3/reviews",
                          "self": "/api/films/public/3"
                      }
                  ],
                  "next": "/api/films/public?pageNo=2"
              }
        "500":
          description: Internal Server Error
...
  "/api/films/public/{filmId}":
    get:
      tags: 
      - "api/films/public/{filmId}"
      operationId: getSinglePublicFilm
      summary: retrieve a specific public film.
      description: A generic user can retrieve a specific public film.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
              example: 
                id: 11
                title: Inception
                owner: 1
                private: false
                reviews: /api/films/public/11/reviews
                self: /api/films/public/11
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    put:
      tags: 
      - "api/films/public/{filmId}"
      security:
        - cookieAuth: []
      operationId: updateSinglePublicFilm
      summary: Update a specific public film of the user.
      description: A logged-in user can update a specific public film that he owns except for it's visibility.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      requestBody:
        description: The film object that must be replaced to the old one
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
            example: 
              title: Inception Remake
              owner: 1
              private: false
        required: true
      responses:
        "200":
          description: Succesfully Updated
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict (The request is not allowed, like a change of visibility)
        "500":
          description: Internal Server Error

    delete:
      tags: 
      - "api/films/public/{filmId}"
      security:
        - cookieAuth: []
      operationId: deleteSinglePublicFilm
      summary: Delete a specific public film of the user.
      description: A logged-in user can delete a specific public film that he owns.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      responses:
        "204":
          description: Successfully Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    Film:
      type: object
      description: The schema for the film data structure
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: The path to the schema that describes the film data structure
        id:
          type: integer
          description: The film's unique identifier
        title:
          type: string
          description: The film's title
        owner:
          type: integer
          description: The id of the user that created the film data structure
        private:
          type: boolean
          default: true
          description: It's true if the owner is the only one that can access the film, false otherwise
        watchDate:
          type: string
          format: date
          description: The date when the owner watched the Film
        rating:
          type: integer
          minimum: 0
          maximum: 10
          description: the rating that the owner has given to the film
        favourite:
          type: boolean
          default: false
          description: It's true if the film is among the owner favourites, false otherwise
        reviews:
          type: string
          format: uri
          description: the reviews that other users have given to the film
        self:
          type: string
          format: uri
          description: URI to access the specific film resource.
      required:
        - title
        - owner
        - private
    Films:
      type: object
      additionalProperties: false
      description: the schema for the films data structure
      properties:
        "$schema":
          type: string
          description: The path to the schema used to validate the films data structure
        totalPages:
          type: integer
          minimum: 1
          description: the total number of pages
        currentPage:
          type: integer
          minimum: 1
          description: the number of the currentPage
        films:
          type: array
          items:
            "$ref": "#/components/schemas/Film"
          description: the list of films on the current page
        next:
          type: string
          format: uri
          description: URI for accessing the next page
      required:
        - totalPages
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid